refactor: rename cacheService.ts to abstractCacheService
