+++ server/src/services/cache/abstractCacheService.ts
@@ -1,6 +1,6 @@
-import { ICache } from "@/viewmodels";
-import { JsonFileHandler } from "../file";
 import winston from "winston";
+import { ICache } from "@viewmodels";
+import { JsonFileHandler } from "../file";
 
 /**
  * @abstract
