+++ server/src/services/cache/index.ts
@@ -1,4 +1,4 @@
 export * from "./projectCacheService";
 export * from "./tagsCacheService";
-export * from './cacheService';
+export * from './abstractCacheService';
 export * from './projectsCache';
